message handler fixing

+++ src/main.ts
@@ -2,11 +2,12 @@ import * as Discord from "discord.js";
 import {handleMessage, addMessageHandler} from "./lib/message";
 import verify from "./passive/verification";
 import {client} from "./client";
-import {handle, isCommand, RESPONSES} from "./lib/command";
 import {setTimeoutCounts} from "./lib/timeout";
+import {handle, isCommand, RESPONSES} from "./lib/command";
 import "./cmd";
 import "./passive/easterEggs";
 import "./passive/log/index";
+//import "./lib/handle";
 
 const statuses = [
     "over the server",
@@ -28,18 +29,18 @@ client.on("ready", () => {
     //initialize timeoutCounts structure on startup
     setTimeoutCounts();
 
+    //ignore bot messages
+    addMessageHandler((message) => message.author.bot);
+
+    //handle commands
+    addMessageHandler(handle);
+
     //automatically update status once every minute
     setInterval(() => {    
         const index:number = Math.floor(Math.random() * (statuses.length - 1));
         client.user.setActivity(statuses[index], {type: "WATCHING"});
     }, 60000);
 });
-
-//ignore bot messages
-addMessageHandler((message) => message.author.bot);
-
-//handle commands
-addMessageHandler(handle);
 
 //verify each member upon entry
 client.on("guildMemberAdd", (member: Discord.GuildMember) => {
@@ -52,19 +53,15 @@ client.on("guildMemberAdd", (member: Discord.GuildMember) => {
 //handle messages appropriately
 client.on("message", handleMessage);
 
-//update command invocations
 client.on("messageUpdate", (old, current) => {
-
-    //ignore old bot messages
+    
+    //ignore bot messages
     if(old.author?.bot) return false;
 
-    //remove the old invocation and response
-    if(isCommand(old) && RESPONSES.has(old)){
-        RESPONSES.get(old)?.delete();
-    }
-
+    //delete old command and update
+    if(isCommand(old) && RESPONSES.has(old)) RESPONSES.get(old)?.delete();
     return handle(current);
-})
+});
 
 //error handling
 //bot will be running on local machine, so just send it to the console
